Support local, no password check JDBC connections
Similar to ODBC, expect "(local)" (with no quotes) to be entered in the
Tableau hostname field.

NOTE username field needs to contain something, the actual contents will
be ignored by the connector/driver and the current operating system
username will be used. NOTE assumes Operating System Authentication is
in effect.